Refactor health-beta to avoid RequiredLoginView conflict (#5924)
