+++ src/js/health-beta/routes.jsx
@@ -1,8 +1,10 @@
 import HealthBetaEnrollment from './containers/HealthBetaEnrollment';
+import Main from './containers/Main';
 
 const routes = {
   path: '/health-beta',
   component: HealthBetaEnrollment,
+  indexRoute: { component: Main },
 };
 
 export default routes;
